Extract printing into separate moduleD
